requirements.txt
fastapi
uvicorn[standard]
langchain
chromadb
python-dotenv
pydantic
openai
anthropic
langchain-community
langchain-openai